requirements.txt
cairocffi
cffi
cmd2
dill
fabric3
funcsigs
jinja2
makesense
matplotlib
networkx
pandas
sh
scipy
